requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   sse-starlette
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==24.2.0
    # via aiohttp
certifi==2024.7.4
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
elastic-transport==8.15.0
    # via elasticsearch
elasticsearch==8.15.0
    # via -r requirements.in
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
fastapi==0.112.1
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   langsmith
    #   openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.3
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.4
    # via -r requirements.in
langchain-core==0.2.5
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.8
    # via -r requirements.in
langchain-text-splitters==0.2.1
    # via langchain
langsmith==0.1.101
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
marshmallow==3.22.0
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.33.0
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.7
    # via langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
    #   pytest
pluggy==1.5.0
    # via pytest
pydantic==2.8.2
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
pytest==8.3.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langsmith
    #   tiktoken
rich==13.7.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via
    #   langchain
    #   langchain-community
sse-starlette==2.1.0
    # via -r requirements.in
starlette==0.38.2
    # via
    #   fastapi
    #   sse-starlette
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.7.0
    # via
    #   -r requirements.in
    #   langchain-openai
tomli==2.0.1
    # via pytest
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.2
    # via
    #   elastic-transport
    #   requests
uvicorn==0.30.6
    # via
    #   -r requirements.in
    #   sse-starlette
yarl==1.9.4
    # via aiohttp
